requirements.txt
Flask>=0.8
Flask-Assets>=0.6.2
Flask-SQLAlchemy>=0.15
Flask-Script>=0.3.1
Flask-WTF>=0.5.4
Flask-Babel>=0.8
psycopg2
unidecode
closure
